# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/balance": {
            "post": {
                "description": "Gets the balance of a user given an user id",
                "produces": [
                    "application/json"
                ],
                "summary": "Gets the balance of a user given an user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id is the unique identitier of the user",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "description": "This endpoint transfers money between two user ids",
                "produces": [
                    "application/json"
                ],
                "summary": "Transfers money between two user ids",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From user Id",
                        "name": "fromUserId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To user Id",
                        "name": "toUserId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Amount to transfer",
                        "name": "amount",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}